feature: ADD focuspoint/openai refine main content.
