+++ src/focuspoint/openai_test.go
@@ -2,20 +2,23 @@ package focuspoint
 
 import (
 	"fmt"
-	"strings"
+	"os"
 	"testing"
+
+	"github.com/joho/godotenv"
+	"github.com/stretchr/testify/assert"
 )
 
-// func makeModule() *FocusPointManager {
-// 	err := godotenv.Load()
-// 	if err != nil {
-// 		fmt.Println("Error loading.")
-// 	}
+func makeModule() *FocusPointManager {
+	err := godotenv.Load()
+	if err != nil {
+		fmt.Println("Error loading.")
+	}
 
-// 	token := os.Getenv("OPEN_AI_KEY")
-// 	fpm := NewFocusPointManager("./img", token)
-// 	return fpm
-// }
+	token := os.Getenv("OPEN_AI_KEY")
+	fpm := NewFocusPointManager("./img", token)
+	return fpm
+}
 
 // func TestCanPostGenerateContent(t *testing.T) {
 // 	// given : 유효한 이미지 + 토큰 있는 경우
@@ -46,26 +49,8 @@ import (
 // 	assert.True(t, generatedContent == "", "예외가 정상적으로 처리되지 않았습니다.")
 // 	assert.Equal(t, err, fmt.Errorf("non token"), "에러 메시지가 정상적으로 처리되지 않았습니다.")
 
-func filterSentences(content string, words []string) string {
-	var filteredSentences []string
-
-	sentences := strings.Split(content, ".")
-OuterLoop:
-	for _, sentence := range sentences {
-		sentence = strings.TrimSpace(sentence)
-		for _, word := range words {
-			if strings.Contains(sentence, word) {
-				continue OuterLoop
-			}
-		}
-		filteredSentences = append(filteredSentences, sentence)
-	}
-
-	result := strings.Join(filteredSentences, ". ")
-	return result
-}
-
 func TestCanRefineMainContent(t *testing.T) {
+	// given : 정상적인 content
 	content := "As an AI developed by OpenAI, I don't have access to external databases to identify specific artwork or artists, and can't comment on the personal life or achievements of this artist without that information. However, I can analyze the image based on art theory and the visible elements.\n" +
 		"This piece of art seems to be a vibrant landscape painting. The medium appears to be acrylic or oil paint on canvas, evident from the texture and the way the light reflects off the surface. The style is somewhat stylized with a slight hint of naïve art due to the simplistic representation of the scenery and the bright, almost uniform colors.\n" +
 		"Influences in the work might include Fauvism for the bold color choices and Asian landscape painting for the terraced fields, which are typically seen in East Asian countries like China and Vietnam. Such fields are often used to grow rice and are sculpted into hilly or mountainous terrain.\n" +
@@ -79,25 +64,13 @@ func TestCanRefineMainContent(t *testing.T) {
 		"\"blue_water\": [[150,300,650,400]]\n" +
 		"}\n" +
 		"```\n"
-
-	// find main content
-	splitContent := strings.Split(content, ":")
-	mainContent := strings.TrimSpace(splitContent[0])
-	// if len(splitContent) > 0 {
-	// 	mainContent := strings.TrimSpace(splitContent[0])
-	// 	fmt.Println(mainContent)
-	// }
-
-	// filter words
-	words := []string{"cannot", "AI", "do not", "can't", "json", "JSON", "{", "Unfortunately", "coordinates", "However", "keyword", "keywords"}
-	filteredContent := filterSentences(mainContent, words)
-	fmt.Println(filteredContent)
-
-	// given : 정상적인 content
-
 	// when : content 추출 요청
+	fpm := makeModule()
+	filteredMainContent, _ := fpm.refineContent(content)
 
 	// then : content
+	fmt.Println(filteredMainContent)
+	assert.True(t, len(filteredMainContent) > 0, "컨텐츠가 정상적으로 생성되지 않았습니다.")
 }
 
 // func TestCanRefindCoordValue(t *testing.T) {
